# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cygnus API",
    "description": "Backend API for the Cygnus project",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/babies": {
      "post": {
        "summary": "Create a baby profile",
        "tags": [
          "Babies"
        ],
        "description": "Creates a new baby profile owned by the authenticated user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "The baby's full name"
                  },
                  "birthDate": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "description": "Date of birth, formatted as YYYY-MM-DD. Cannot be in the future"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "MALE",
                      "FEMALE"
                    ]
                  },
                  "bloodType": {
                    "type": "string",
                    "enum": [
                      "A+",
                      "A-",
                      "B+",
                      "B-",
                      "AB+",
                      "AB-",
                      "O+",
                      "O-"
                    ]
                  },
                  "allergies": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    },
                    "description": "Known allergies, e.g. lactose, penicillin"
                  },
                  "healthPlanName": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Health insurance plan, e.g. Unimed"
                  },
                  "healthPlanNumber": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Member number printed on the insurance card — what a clinic desk asks for"
                  },
                  "avatarUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "avatarColor": {
                    "type": "string",
                    "pattern": "^#[0-9A-Fa-f]{6}$",
                    "description": "Hex color used as the avatar's border, e.g. #2A9D8F"
                  }
                },
                "required": [
                  "name",
                  "birthDate",
                  "gender"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "A baby profile owned by the authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "userId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "birthDate": {
                      "type": "string",
                      "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                    },
                    "gender": {
                      "type": "string",
                      "enum": [
                        "MALE",
                        "FEMALE"
                      ]
                    },
                    "bloodType": {
                      "type": "string",
                      "nullable": true
                    },
                    "allergies": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "healthPlanName": {
                      "type": "string",
                      "nullable": true
                    },
                    "healthPlanNumber": {
                      "type": "string",
                      "nullable": true
                    },
                    "avatarUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "avatarColor": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "userId",
                    "name",
                    "birthDate",
                    "gender",
                    "bloodType",
                    "allergies",
                    "healthPlanName",
                    "healthPlanNumber",
                    "avatarUrl",
                    "avatarColor",
                    "createdAt"
                  ],
                  "additionalProperties": false,
                  "description": "A baby profile owned by the authenticated user"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          },
          "401": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          },
          "500": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List the authenticated user's baby profiles",
        "tags": [
          "Babies"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "userId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string"
                      },
                      "birthDate": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      },
                      "gender": {
                        "type": "string",
                        "enum": [
                          "MALE",
                          "FEMALE"
                        ]
                      },
                      "bloodType": {
                        "type": "string",
                        "nullable": true
                      },
                      "allergies": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "healthPlanName": {
                        "type": "string",
                        "nullable": true
                      },
                      "healthPlanNumber": {
                        "type": "string",
                        "nullable": true
                      },
                      "avatarUrl": {
                        "type": "string",
                        "nullable": true
                      },
                      "avatarColor": {
                        "type": "string",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    },
                    "required": [
                      "id",
                      "userId",
                      "name",
                      "birthDate",
                      "gender",
                      "bloodType",
                      "allergies",
                      "healthPlanName",
                      "healthPlanNumber",
                      "avatarUrl",
                      "avatarColor",
                      "createdAt"
                    ],
                    "additionalProperties": false,
                    "description": "A baby profile owned by the authenticated user"
                  }
                }
              }
            }
          },
          "401": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          },
          "500": {
            "description": "The request could not be completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "message"
                  ],
                  "additionalProperties": false,
                  "description": "The request could not be completed"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}